requirements.txt
django
django-environ
django-cors-headers
celery
redis
channels
uvicorn[standard]
channels_redis